Fix bug in getter
There was a blind spot in the test suite wherein I was generating the
exactly right chunks of bytestrings, then in the test consuming them as
a lazy bytestring of that same size. I was not using the internal
deserialization function that properly handled leftovers, but the tests
still passed. I exposed the bug by randomly rechunking the stream before
parsing it. The solution was to actually use the internal parsing
function.